requirements.txt
appnope==0.1.0
asgiref==3.2.3
attrs==19.3.0
autobahn==19.11.2
Automat==0.8.0
backports.shutil-get-terminal-size==1.0.0
certifi==2018.10.15
cffi==1.13.2
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
decorator==4.3.0
Django==2.2.9
django-cors-headers==2.4.0
django-extensions==2.1.3
django-filer==1.6.0
django-mptt==0.8.7
django-polymorphic==2.0.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
easy-thumbnails==2.5
hyperlink==19.0.0
idna==2.7
importlib-metadata==1.3.0
incremental==17.5.0
ipdb==0.11
ipython==5.8.0
ipython-genutils==0.2.0
more-itertools==8.0.2
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.2.0
prompt-toolkit==1.0.15
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pyOpenSSL==19.1.0
pytz==2018.7
requests==2.20.1
scandir==1.9.0
service-identity==18.1.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.3.0
traitlets==4.3.2
Twisted==19.10.0
txaio==18.8.1
typing==3.6.6
Unidecode==0.4.21
urllib3==1.24.2
wcwidth==0.1.7
zipp==0.6.0
zope.interface==4.7.1
